requirements.txt
datasets==3.5.0
faiss-cpu==1.10.0
gitpython==3.1.44
langgraph==0.3.25
langchain_community==0.3.21
langchain-openai==0.3.12
rank_bm25==0.2.2
requests==2.32.3